+++ nix/default.nix
@@ -5,7 +5,7 @@
 {
   self,
   nixpkgs,
-  rs-harbor,
+  harbor-rs,
   rust-overlay,
   treefmt-nix,
   git-hooks,
@@ -14,7 +14,7 @@
   ...
 }:
 nix-manager-core.lib.mkManagerOutputs {
-  inherit self nixpkgs rs-harbor rust-overlay treefmt-nix git-hooks;
+  inherit self nixpkgs harbor-rs rust-overlay treefmt-nix git-hooks;
   crateName = "dns-manager";
   srcDir = ../.;
 
@@ -91,10 +91,10 @@ nix-manager-core.lib.mkManagerOutputs {
       then {}
       else let
         pkgs = pkgsFor system;
-        toolchain = rs-harbor.lib.mkToolchain {inherit pkgs; toolchainProfile = "nightly";};
-        cross = rs-harbor.lib.mkCross {inherit pkgs system;};
+        toolchain = harbor-rs.lib.mkToolchain {inherit pkgs; toolchainProfile = "nightly";};
+        cross = harbor-rs.lib.mkCross {inherit pkgs system;};
       in {
-        docs = rs-harbor.lib.mkDocsShell {
+        docs = harbor-rs.lib.mkDocsShell {
           inherit pkgs cross;
           inherit (toolchain) craneLib;
           packages = with pkgs; [